Splitting of sources and mappings in seperated packages.
The sources (OCS computers and OCS SNMP devices) are moved to seperated
classes in the source package. A replay version for each source was
added, allowing to replay a dump containing the SOP responses.

The mapping was splitted into a default mapper for each source and a
script mapper. These are located in the mapping package.

The mode handling was moved to seperated classes now located in the
driver package.

Further, the configuration loading was extracted into a config manager
allowing a base configuration and a configuration per OCS instance in a
sub-folder. The name of the sub-fooker must match the name of the
instance. The config loading is now done using Apacha Commons
Configuration

The BSF scripting engine was replaced by the Java Scripting API.

The com.sun web server implementation was replaced by jetty.