Add CircleCI build environment
Build and test the code with CircleCI. Create and push docker image to DockerHub introducing CI/CD workflows.
Added smoke test phase where container is started and smoke tests run against:

- landing page which redirects to the documentation
 - myRouter example requisition
 - myServer example requisition

The output of the example requisitions is persisted as artifact of the test phase.

Added a publishing workflow for latest, stable and development image to DockerHub based on a branch name filter.
Refactored release versions to following scheme:

- master: is merge target for next release branch, tagged and published as bleeding
- release-*: release branches, tagged and published as specific version number and latest